+++ src/core/Constants.h
@@ -1,0 +1,5 @@
+#pragma once 
+
+inline constexpr int TILE_SIZE = 32;
+
+inline const char* MAP_PATH_1 = "assets/maps/map2.json";
