+++ src/repo/map/TiledMapLoader.h
@@ -1,0 +1,16 @@
+#pragma once
+#include <string> 
+#include "MapTypes.h"
+
+class Map;
+class Item;
+
+class TiledMapLoader {
+private:
+    static SDL_Texture* loadTexture(const std::string& path, SDL_Renderer* renderer);
+
+public:
+    static MapData loadMap(const std::string& path, SDL_Renderer* renderer);
+    static std::vector<Item> LoadItemsFromMap(const std::string& path, MapData& mapData);
+
+};
